requirements.txt
optimum[onnxruntime]
transformers
os
pandas
scikit-learn
matplotlib
seaborn
torch
deep-translator
dvc
pyyaml
mlflow
boto3
awscli 